Refactor variant to GetRootResponse and cache it
